+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/coupon/CouponRepositoryImpl.java
@@ -1,9 +1,17 @@
 package kr.hhplus.be.server.infrastructure.jpa.coupon;
 
+import static kr.hhplus.be.server.domain.coupon.QIssuedCoupon.issuedCoupon;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import java.util.List;
 import java.util.Optional;
 import kr.hhplus.be.server.domain.coupon.Coupon;
+import kr.hhplus.be.server.domain.coupon.IssuedCoupon;
 import kr.hhplus.be.server.domain.coupon.repository.CouponRepository;
+import kr.hhplus.be.server.infrastructure.jpa.coupon.issuedcoupon.IssuedCouponJpaRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -20,5 +28,36 @@ public class CouponRepositoryImpl implements CouponRepository {
     @Override
     public Optional<Coupon> findById(long couponId) {
         return couponJpaRepository.findById(couponId);
+    }
+
+    private final IssuedCouponJpaRepository issuedCouponJpaRepository;
+    private final JPAQueryFactory queryFactory;
+
+    @Override
+    public IssuedCoupon saveIssuedCoupon(IssuedCoupon issuedCoupon) {
+        return issuedCouponJpaRepository.save(issuedCoupon);
+    }
+
+    @Override
+    public Page<IssuedCoupon> selectIssuedCouponList(long userId, Pageable pageable) {
+        List<IssuedCoupon> issuedCoupons = queryFactory
+            .selectFrom(issuedCoupon)
+            .where(issuedCoupon.user.id.eq(userId))
+            .orderBy(issuedCoupon.createdAt.desc())
+            .offset(pageable.getOffset())
+            .limit(pageable.getPageSize())
+            .fetch();
+
+        long total = queryFactory
+            .selectFrom(issuedCoupon)
+            .where(issuedCoupon.user.id.eq(userId))
+            .fetchCount();
+
+        return new PageImpl<>(issuedCoupons, pageable, total);
+    }
+
+    @Override
+    public Optional<IssuedCoupon> findIssuedCouponById(long issuedCouponId) {
+        return issuedCouponJpaRepository.findById(issuedCouponId);
     }
 }
